# Concurrent fiber solver