--- a/otherlibs/stdune/src/hashtbl_intf.ml
+++ b/otherlibs/stdune/src/hashtbl_intf.ml
@@ -21,4 +21,5 @@ module type S = sig
   val to_dyn : ('v -> Dyn.t) -> 'v t -> Dyn.t
   val filteri_inplace : 'a t -> f:(key:key -> data:'a -> bool) -> unit
   val length : _ t -> int
+  val to_list : 'a t -> (key * 'a) list
 end