Switch to parallel iteration from sequential iteration. This required switching to a fiber aware cache.

This gives a small but nice boost:

Before:
```
Benchmark 1: ~/github/ocaml/dune/_build/default/bin/main.exe pkg lock
  Time (mean ¬± œÉ):      1.962 s ¬±  0.069 s    [User: 1.285 s, System: 0.444 s]
  Range (min ‚Ä¶ max):    1.882 s ‚Ä¶  2.086 s    10 runs
```

After:
```
Benchmark 1: ~/github/ocaml/dune/_build/default/bin/main.exe pkg lock
  Time (mean ¬± œÉ):      1.752 s ¬±  0.315 s    [User: 1.284 s, System: 0.464 s]
  Range (min ‚Ä¶ max):    1.535 s ‚Ä¶  2.606 s    10 runs
```